+++ R/zzz.R
@@ -11,7 +11,8 @@
   utils::globalVariables(c(
     "vec_is_vector_workaround",
     "vec_type2_workaround",
-    "vec_cast_workaround"
+    "vec_cast_workaround",
+    "vec_slice_workaround"
   ))
 
   # Work around "invalid generic in usemethod" error on R 3.1
@@ -22,7 +23,9 @@
       vec_type2_workaround = vec_type2_dispatch,
       vec_type2_dispatch = function(...) vec_type2_workaround(...),
       vec_cast_workaround = vec_cast_dispatch,
-      vec_cast_dispatch = function(...) vec_cast_workaround(...)
+      vec_cast_dispatch = function(...) vec_cast_workaround(...),
+      vec_slice_workaround = vec_slice_dispatch,
+      vec_slice_dispatch = function(...) vec_slice_workaround(...)
     )
   }