+++ src/utils.h
@@ -4,7 +4,9 @@
 
 bool is_bool(SEXP x);
 
-SEXP vctrs_dispatch3(SEXP fn, SEXP x, SEXP y);
+SEXP vctrs_dispatch2(SEXP fn_sym, SEXP fn,
+                     SEXP x_sym, SEXP x,
+                     SEXP y_sym, SEXP y);
 
 R_len_t r_lgl_sum(SEXP lgl, bool na_true);
 SEXP r_lgl_which(SEXP x, bool na_true);
@@ -15,6 +17,15 @@ void r_int_fill(SEXP x, int value);
 void r_int_fill_seq(SEXP x, int start);
 
 bool r_int_any_na(SEXP x);
+
+SEXP r_new_environment(SEXP parent, R_len_t size);
+
+
+extern SEXP syms_i;
+extern SEXP syms_x;
+extern SEXP syms_y;
+extern SEXP syms_to;
+extern SEXP syms_dots;
 
 
 #endif