+++ NEWS.md
@@ -1,5 +1,7 @@
 # vctrs 0.1.0.9000
 
+* `vec_slice()` is now a generic. Its default method calls `[`.
+
 * `vec_size()` now works with positive short row names. This fixes
   issues with data frames created with jsonlite (#220).